Add .NET Core as test target; Migrate test projects to cleaner csproj syntax; Fix .NET Core specific bugs
