+++ Sources/Core/Microsoft.StreamProcessing/Microsoft.StreamProcessing.csproj
@@ -15,6 +15,7 @@
     <PackageReference Include="Microsoft.CodeAnalysis.Scripting" Version="3.1.0" />
     <PackageReference Include="System.Diagnostics.Contracts" Version="4.3.0" />
     <PackageReference Include="System.Diagnostics.Process" Version="4.3.0" />
+    <PackageReference Include="System.Linq.Expressions" Version="4.3.0" />
     <PackageReference Include="System.Runtime.Serialization.Primitives" Version="4.3.0" />
     <PackageReference Include="System.Threading.Thread" Version="4.3.0" />
     <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
