+++ Sources/Core/Microsoft.StreamProcessing/StreamProcessing.nuspec
@@ -19,6 +19,7 @@
         <dependency id="NETStandard.Library" version="2.0.0" />
         <dependency id="System.Diagnostics.Contracts" version="4.3.0" />
         <dependency id="System.Diagnostics.Process" version="4.3.0" />
+        <dependency id="System.Linq.Expressions" version="4.3.0" />
         <dependency id="System.Runtime.Loader" version="4.3.0" />
         <dependency id="System.Runtime.Serialization.Primitives" version="4.3.0" />
         <dependency id="System.Threading.Thread" version="4.3.0" />
