+++ Sources/Test/SimpleTesting/CheckpointRestoreTests.tt
@@ -29,26 +29,6 @@ namespace SimpleTesting
         [DataMember]
         public ulong ActivityCount;
     }
-
-    public abstract class TestWithConfigSettingsWithoutMemoryLeakDetection
-    {
-        private IDisposable confmod = null;
-        private readonly ConfigModifier modifier;
-
-        protected TestWithConfigSettingsWithoutMemoryLeakDetection() : this(new ConfigModifier()) { }
-
-        internal TestWithConfigSettingsWithoutMemoryLeakDetection(ConfigModifier modifier) => this.modifier = modifier;
-
-        [TestInitialize]
-        public virtual void Setup() => this.confmod = this.modifier.Modify();
-
-        [TestCleanup]
-        public virtual void TearDown()
-        {
-            var cm = System.Threading.Interlocked.Exchange(ref this.confmod, null);
-            if (cm != null) { cm.Dispose(); }
-        }
-    }
 <#
 foreach (var orientation in new [] { "Row", "Columnar" })
 foreach (var batch in new [] { string.Empty, "SmallBatch" })
