+++ Sources/Test/SimpleTesting/SimpleTests.tt
@@ -538,11 +538,11 @@ foreach (var disorderPolicy in (new [] { string.Empty, "Floating" }))
 <# if (string.IsNullOrEmpty(fusionOption)) { #>
                 .ShiftEventLifetime(0)
 <# } #>
-                .SelectMany(e => e.field2.ToString().Split('.').Select(s => s.Length))
+                .SelectMany(e => e.field2.ToString().Split('.', StringSplitOptions.None).Select(s => s.Length))
                 .ToPayloadEnumerable()
                 ;
             var linqResult = input
-                .SelectMany(e => e.field2.ToString().Split('.').Select(s => s.Length))
+                .SelectMany(e => e.field2.ToString().Split('.', StringSplitOptions.None).Select(s => s.Length))
                 ;
             Assert.IsTrue(linqResult.SequenceEqual(streamResult));
         }
