--- Sources/Test/TrillPerf/AssemblyInfo.cs
@@ -1,9 +1,0 @@
-﻿// *********************************************************************
-// Copyright (c) Microsoft Corporation.  All rights reserved.
-// Licensed under the MIT License
-// *********************************************************************
-[assembly: System.Reflection.AssemblyTitle("TrillPerf")]
-[assembly: System.Reflection.AssemblyDescription("")]
-[assembly: System.Reflection.AssemblyVersion("1.0.0.0")]
-
-[assembly: System.Runtime.InteropServices.ComVisible(false)]
