+++ Sources/Test/TrillPerf/PerformanceTesting.csproj
@@ -1,69 +1,18 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="4.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
   <PropertyGroup>
-    <ProjectGuid>{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>PerformanceTesting</RootNamespace>
-    <AssemblyName>PerformanceTesting</AssemblyName>
-    <FileAlignment>512</FileAlignment>
-    <NuGetPackageImportStamp>
-    </NuGetPackageImportStamp>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
+    <TargetFrameworks>net472;netcoreapp3.1</TargetFrameworks>
+    <Platforms>AnyCPU</Platforms>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
-    <DebugSymbols>true</DebugSymbols>
-    <OutputPath>bin\Debug\</OutputPath>
-    <DefineConstants>;DEBUG;TRACE;CODE_ANALYSIS;BINARYMERGE;COLUMNAR;BITVECTOR;MULTICORE_GA</DefineConstants>
-    <DocumentationFile>F:\dev\Trill\Sources\Test\..\..\bintest\x64\Debug\TrillPerf.xml</DocumentationFile>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <NoWarn>1591</NoWarn>
-    <DebugType>full</DebugType>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <ErrorReport>prompt</ErrorReport>
-    <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
+
+  <ItemGroup>
+    <PackageReference Include="System.Reactive" Version="4.3.2" />
+    <PackageReference Include="System.Reactive.Linq" Version="4.1.5" />
+  </ItemGroup>
+
+  <PropertyGroup>
+    <TransformOnBuild>true</TransformOnBuild>
   </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'">
-    <DebugSymbols>true</DebugSymbols>
-    <OutputPath>bin\Release\</OutputPath>
-    <DefineConstants>;TRACE;BINARYMERGE;COLUMNAR;BITVECTOR;MULTICORE_GA</DefineConstants>
-    <DocumentationFile>F:\dev\Trill\Sources\Test\..\..\bintest\x64\Release\TrillPerf.xml</DocumentationFile>
-    <Optimize>true</Optimize>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <NoWarn>1591</NoWarn>
-    <DebugType>pdbonly</DebugType>
-    <PlatformTarget>AnyCPU</PlatformTarget>
-    <ErrorReport>prompt</ErrorReport>
-    <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Include="IPerfTestState.cs" />
-    <Compile Include="PerfTestAttribute.cs" />
-    <Compile Include="PerfTestManager.cs" />
-    <Compile Include="Program.cs" />
-    <Compile Include="AssemblyInfo.cs" />
-    <Compile Include="Streamables\EquiJoinStreamablePerfTest.cs" />
-    <Compile Include="YahooBenchmark\DataGenerator.cs" />
-    <Compile Include="YahooBenchmark\PrimaryQuery.cs" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\app.config">
-      <Link>app.config</Link>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\Core\Microsoft.StreamProcessing\Microsoft.StreamProcessing.csproj">
-      <Project>{29810b58-a929-4127-83df-5004d94c8641}</Project>
-      <Name>Microsoft.StreamProcessing</Name>
-    </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="System.Reactive.Linq">
-      <Version>4.1.5</Version>
-    </PackageReference>
-  </ItemGroup>
-  <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+
 </Project>
