--- Sources/Test/TrillTest/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="System.Runtime" version="4.1.0" targetFramework="net46" />
-</packages>
