+++ Sources/Trill.sln
@@ -43,20 +43,20 @@ Global
 		{3A2EB89B-70C7-4516-BB40-13248D5D2AB9}.Release|x64.Build.0 = Release|x64
 		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Debug|x64.ActiveCfg = Debug|x64
-		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Debug|x64.Build.0 = Debug|x64
+		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Debug|x64.Build.0 = Debug|Any CPU
 		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Release|x64.ActiveCfg = Release|x64
-		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Release|x64.Build.0 = Release|x64
+		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Release|x64.ActiveCfg = Release|Any CPU
+		{F88FBCF8-C33B-4089-8998-19E9CF901B99}.Release|x64.Build.0 = Release|Any CPU
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Debug|x64.ActiveCfg = Debug|x64
-		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Debug|x64.Build.0 = Debug|x64
+		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Debug|x64.Build.0 = Debug|Any CPU
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|x64.ActiveCfg = Release|x64
-		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|x64.Build.0 = Release|x64
+		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|x64.ActiveCfg = Release|Any CPU
+		{5DD350C9-C6A8-4EBF-A192-BAEFD26013FB}.Release|x64.Build.0 = Release|Any CPU
 		{E225184C-B9AE-4F22-9A03-96C5346E3DCF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E225184C-B9AE-4F22-9A03-96C5346E3DCF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E225184C-B9AE-4F22-9A03-96C5346E3DCF}.Debug|x64.ActiveCfg = Debug|x64
